Include New Transport typeface
This typeface is included in the govuk_template project, which we're
not including as a dependency here because big screen view will never
want the header or footer.

Instead, I've duplicated the web fonts and the fonts.scss file here.
This isn't a nice solution but I think it's the best one available
to us right now.