Make build destination configurable
To deploy this application to GOV.UK infrastructure, we want to be
able to serve this application from a `public/` directory.

This commit moves the config from a JavaScript object to a JSON
file but doesn't change any of the config. This JavaScript file
will be overwritten at deploy time when deploying to GOV.UK.

Paired with @timmow on this.