Remove Travis Functional Tests
Have left an empty Gulp Travis function as this is still looked for and is required for the build
to pass.  Function tests removed as these have been broken in the master branch for a while, and do
not appear to be easily fixable.  This needs to be looked into at a later date.